# docs/HOMEZ_ONCHANNEL_OPENAPI_SPEC_20260908.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "온채널 OpenAPI",
        "version": "0.1",
        "x-name": "openapi"
    },
    "paths": {
...
        "/openapi/seller/product": {
            "get": {
                "tags": [
                    "판매사 - 상품"
                ],
                "summary": "상품 리스트",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "페이지 number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "상품상태(1: 정상판매, 2: 단종, 3: 판매중단, 4: 일시품절, 5: 품절)",
                        "required": false,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/product.status"
                                }
                            ]
                        },
                        "example": null
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "리스트 표시 수(max: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/success"
                                        },
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "description": "상품ID",
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "channel": {
                                                                        "$ref": "#/components/schemas/product.channel"
                                                                    },
                                                                    "product_code": {
                                                                        "description": "상품코드",
                                                                        "type": "string",
                                                                        "example": "CH1894996"
                                                                    },
                                                                    "name": {
                                                                        "description": "상품명",
                                                                        "type": "string",
                                                                        "example": "테스트 상품명"
                                                                    },
                                                                    "status": {
                                                                        "$ref": "#/components/schemas/product.status"
                                                                    },
                                                                    "created_at": {
                                                                        "description": "등록일",
                                                                        "type": "string",
                                                                        "example": "2022-08-09 15:48:26"
                                                                    },
                                                                    "updated_at": {
                                                                        "description": "수정일",
                                                                        "type": "string",
                                                                        "example": "2026-01-05 15:14:39"
                                                                    },
                                                                    "options": {
                                                                        "description": "상품 옵션",
                                                                        "type": "array",
                                                                        "items": {
                                                                            "properties": {
                                                                                "id": {
                                                                                    "description": "옵션ID",
                                                                                    "type": "integer",
                                                                                    "example": 22431922
                                                                                },
                                                                                "status": {
                                                                                    "$ref": "#/components/schemas/option.status"
                                                                                },
                                                                                "name": {
                                                                                    "description": "옵션명",
                                                                                    "type": "string",
                                                                                    "example": "청색 230"
                                                                                },
                                                                                "qty": {
                                                                                    "description": "재고수량",
                                                                                    "type": "integer",
                                                                                    "example": 500
                                                                                }
                                                                            },
                                                                            "type": "object"
                                                                        }
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "total_cnt": {
                                                            "type": "integer",
                                                            "example": 57
                                                        },
                                                        "page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "page_size": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 57
                                                        },
                                                        "_pager": {
                                                            "properties": {
                                                                "has_more": {
                                                                    "type": "boolean",
                                                                    "example": true
                                                                },
                                                                "next_url": {
                                                                    "type": "string",
                                                                    "example": "http://localhost:8084/openapi/supplier/product?page=2"
                                                                },
                                                                "prev_url": {
                                                                    "nullable": true,
                                                                    "type": "string",
                                                                    "example": null
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "success": {
                "title": "통신 성공 Scheme",
                "description": "통신 성공 Scheme",
                "required": [
                    "status",
                    "meta",
                    "result"
                ],
                "properties": {
                    "status": {
                        "description": "상태코드",
                        "type": "integer",
                        "example": 200
                    },
                    "meta": {
                        "properties": {
                            "timestamp": {
                                "description": "통신 시간",
                                "type": "string",
                                "example": "2026-01-01 00:00:00"
                            },
                            "api_type": {
                                "description": "api 종류",
                                "type": "string",
                                "example": "openapi"
                            }
                        },
                        "type": "object"
                    },
                    "result": {
                        "description": "결과",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "error": {
                "title": "통신 실패 Scheme",
                "description": "통신 실패 Scheme",
                "required": [
                    "status",
                    "meta",
                    "error"
                ],
                "properties": {
                    "status": {
                        "description": "상태코드",
                        "type": "integer",
                        "example": 401
                    },
                    "meta": {
                        "properties": {
                            "timestamp": {
                                "description": "통신 시간",
                                "type": "string",
                                "example": "2026-01-01 00:00:00"
                            },
                            "api_type": {
                                "description": "api 종류",
                                "type": "string",
                                "example": "openapi"
                            }
                        },
                        "type": "object"
                    },
                    "error": {
                        "properties": {
                            "code": {
                                "description": "에러코드",
                                "type": "integer",
                                "example": 401
                            },
                            "message": {
                                "description": "에러 내용",
                                "type": "string",
                                "example": "잘못 된 접근입니다."
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "product.channel": {
                "title": "상품채널(1, 3만 등록/수정 가능)",
                "description": "1: 가격자율, 2: 유통과학센터, 3: 가격준수, 4: (구)공동구매, 5: 폐쇄몰, 6: 종합몰, 7: 특판/판촉물, 8: 해외마켓, 9: 폐쇄몰(한정), 10: 전문몰, 11: 박스단위, 12: OEM단위, 13: 프랜차이즈, 14: 엄격준수, 15: 가격자율, 16: 가격준수, 17: 이벤트상품, 21: 해외직배송, 22: 해외채널, 24: 재고채널, 25: 단독상품관, 26: 틈새상품관, 27: 소상공인, 28: 사입상품관, 30: 중국도매상품관, 80: 감자마켓, 81: 오프라인, 82: 공동구매, 97: 마켓 준수, 98: 마켓 자율, 99: 제휴관",
                "type": "integer",
                "enum": [
                    1,
                    2,
                    3,
                    4,
                    5,
                    6,
                    7,
                    8,
                    9,
                    10,
                    11,
                    12,
                    13,
                    14,
                    15,
                    16,
                    17,
                    21,
                    22,
                    24,
                    25,
                    26,
                    27,
                    28,
                    30,
                    80,
                    81,
                    82,
                    97,
                    98,
                    99
                ],
                "example": 1
            },
            "product.status": {
                "title": "상품상태",
                "description": "1: 정상판매, 2: 단종, 3: 판매중단, 4: 일시품절, 5: 품절",
                "type": "integer",
                "enum": [
                    1,
                    2,
                    3,
                    4,
                    5
                ],
                "example": 1
            },
            "option.status": {
                "title": "옵션상태",
                "description": "1: 정상(재입고), 2: 단종, 3: 품절",
                "type": "integer",
                "enum": [
                    1,
                    2,
                    3
                ],
                "example": 1
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}